Updat fish if some apps doesn't installed, prepare for Silverblue switch

+++ .config/fish/functions/br.fish
@@ -1,12 +1,16 @@
 function br
-    set -l cmd_file (mktemp)
-    if broot --outcmd $cmd_file $argv
-        read --local --null cmd < $cmd_file
-        rm -f $cmd_file
-        eval $cmd
+    if command -q broot
+        set -l cmd_file (mktemp)
+        if broot --outcmd $cmd_file $argv
+            read --local --null cmd < $cmd_file
+            rm -f $cmd_file
+            eval $cmd
+        else
+            set -l code $status
+            rm -f $cmd_file
+            return $code
+        end
     else
-        set -l code $status
-        rm -f $cmd_file
-        return $code
+        echo "broot doesn't installed"
     end
 end
